Remover propiedad para tipo de factura de los modelos de request (DTE-01 y DTE-03); Finalización de #28
