+++ Source/InvoizR/InvoizR.Application/Features/Invoices/Commands/CreateDte03InvoiceOWCommandHandler.cs
@@ -4,6 +4,7 @@ using InvoizR.Clients.DataContracts.Common;
 using InvoizR.Domain.Entities;
 using InvoizR.Domain.Enums;
 using InvoizR.Domain.Exceptions;
+using InvoizR.SharedKernel.Mh.FeCcf;
 using MediatR;
 using Microsoft.Extensions.Logging;
 
@@ -22,7 +23,7 @@ public class CreateDte03InvoiceOWCommandHandler : IRequestHandler<CreateDte03Inv
 
     public async Task<CreatedResponse<long?>> Handle(CreateDte03InvoiceOWCommand request, CancellationToken cancellationToken)
     {
-        _ = await _dbContext.GetCurrentInvoiceTypeAsync(request.InvoiceTypeId, ct: cancellationToken) ?? throw new InvalidCurrentInvoiceTypeException();
+        _ = await _dbContext.GetCurrentInvoiceTypeAsync(FeCcfv3.TypeId, ct: cancellationToken) ?? throw new InvalidCurrentInvoiceTypeException();
 
         var txn = await _dbContext.Database.BeginTransactionAsync(cancellationToken);
 
@@ -44,7 +45,7 @@ public class CreateDte03InvoiceOWCommandHandler : IRequestHandler<CreateDte03Inv
                 CustomerPhone = request.Customer.Phone,
                 CustomerEmail = request.Customer.Email,
                 CustomerLastUpdated = DateTime.Now,
-                InvoiceTypeId = request.InvoiceTypeId,
+                InvoiceTypeId = FeCcfv3.TypeId,
                 InvoiceNumber = request.InvoiceNumber,
                 InvoiceDate = request.InvoiceDate,
                 InvoiceTotal = request.InvoiceTotal,
@@ -70,7 +71,7 @@ public class CreateDte03InvoiceOWCommandHandler : IRequestHandler<CreateDte03Inv
         {
             await txn.RollbackAsync(cancellationToken);
 
-            _logger.LogCritical(ex, "There was an error on Create Dte03 Invoice");
+            _logger.LogCritical(ex, "There was an error on Create DTE-03 Invoice in OW processing");
 
             return new();
         }
