+++ Source/InvoizR/InvoizR.Client.CapsuleCorp/Program.cs
@@ -1,14 +1,14 @@
 ﻿using InvoizR.Client.CapsuleCorp;
 using InvoizR.Client.CapsuleCorp.Data;
-using InvoizR.Client.CapsuleCorp.Helpers;
 using InvoizR.Client.CapsuleCorp.Mocks;
 using InvoizR.Clients;
 using InvoizR.Clients.Contracts;
 using InvoizR.Clients.DataContracts;
-using InvoizR.SharedKernel;
 using InvoizR.SharedKernel.Mh.FeCcf;
 using InvoizR.SharedKernel.Mh.FeFc;
 using Microsoft.Extensions.Options;
+
+const string BillingEndpoint = "https://localhost:13880";
 
 var clientArgs = new ClientArgs(args);
 
@@ -56,7 +56,7 @@ if (clientArgs.ShowCatalog)
 
 InvoizRClient client = new(Options.Create(new InvoizRClientSettings
 {
-    Endpoint = "https://localhost:13880"
+    Endpoint = BillingEndpoint
 }));
 
 if (clientArgs.Seed)
@@ -153,7 +153,7 @@ if (clientArgs.Mock)
 
 static async Task MockRtDte01(IInvoizRClient client)
 {
-    var req = InvoiceMocker.MockRtDte01();
+    var req = DTE01.MockRtDte01();
 
     Console.WriteLine($" Syncing invoice '{req.InvoiceNumber}', {req.InvoiceTotal:C2} in RT processing...");
 
@@ -165,7 +165,7 @@ static async Task MockRtDte01(IInvoizRClient client)
 
 static async Task MockRtDte03(IInvoizRClient client)
 {
-    var req = InvoiceMocker.MockRtDte03();
+    var req = DTE03.MockRtDte03();
 
     Console.WriteLine($" Syncing invoice '{req.InvoiceNumber}', {req.InvoiceTotal:C2} in RT processing...");
 
@@ -177,7 +177,7 @@ static async Task MockRtDte03(IInvoizRClient client)
 
 static async Task MockOwDte01(IInvoizRClient client)
 {
-    var req = InvoiceMocker.MockOwDte01();
+    var req = DTE01.MockOwDte01();
 
     Console.WriteLine($" Syncing invoice '{req.InvoiceNumber}', {req.InvoiceTotal:C2} in OW processing...");
 
@@ -189,7 +189,7 @@ static async Task MockOwDte01(IInvoizRClient client)
 
 static async Task MockOwDte03(IInvoizRClient client)
 {
-    var req = InvoiceMocker.MockOwDte03();
+    var req = DTE03.MockOwDte03();
 
     Console.WriteLine($" Syncing invoice '{req.InvoiceNumber}', {req.InvoiceTotal:C2} in OW processing...");
 
@@ -197,251 +197,4 @@ static async Task MockOwDte03(IInvoizRClient client)
 
     Console.WriteLine($"  Generated ID: '{response.Id}'");
     Console.WriteLine();
-}
-
-class InvoiceMocker
-{
-    public static CreateDte01InvoiceOWCommand MockOwDte01()
-    {
-        var data = new
-        {
-            Branches = Db.Branches.ToList(),
-            Products = Db.Products.ToList(),
-            Persons = Db.Cards.Where(item => item.CardTypeId == (short)CardType.Person).ToList(),
-            WalkIns = Db.Cards.Where(item => item.CardTypeId == (short)CardType.WalkIn).ToList()
-        };
-
-        var lines = Random.Shared.Next(1, data.Products.Count);
-        var invoiceLines = new List<InvoiceLine>();
-
-        for (var i = 0; i < lines; i++)
-        {
-            var randomIndex = Random.Shared.Next(0, lines - 1);
-            var product = data.Products[randomIndex];
-            if (invoiceLines.Any(item => item.Code == product.Code))
-                continue;
-
-            invoiceLines.Add(new(product.Code, product.Description, (double)product.Price, 1));
-        }
-
-        var total = Math.Round(invoiceLines.Sum(item => item.Total), 2);
-
-        var req = new CreateDte01InvoiceOWCommand
-        {
-            PosId = (short)Random.Shared.Next(1, data.Branches.Count),
-            InvoiceTypeId = FeFcv1.TypeId,
-            InvoiceNumber = Convert.ToInt64($"{DateTime.Now:MMddhhmmss}"),
-            InvoiceDate = DateTime.Now,
-            InvoiceTotal = (decimal)total,
-            Lines = lines,
-        };
-
-        if (req.InvoiceTotal > CreateDte01InvoiceCommand.MaxAmountForAnonymousCustomers)
-        {
-            var persons = data.Persons.ToList();
-            var randomCard = persons[Random.Shared.Next(0, persons.Count - 1)];
-
-            req.Customer.Id = randomCard.Id;
-            req.Customer.DocumentTypeId = randomCard.DocumentTypeId;
-            req.Customer.DocumentNumber = randomCard.DocumentNumber;
-            req.Customer.WtId = randomCard.WtId;
-            req.Customer.Name = randomCard.Name;
-            req.Customer.CountryId = randomCard.CountryId;
-            req.Customer.CountryLevelId = randomCard.CountryLevelId;
-            req.Customer.Address = randomCard.Address;
-            req.Customer.Phone = randomCard.Phone;
-            req.Customer.Email = randomCard.Email;
-        }
-        else
-        {
-            var card = data.WalkIns.First();
-
-            req.Customer.Id = card.Id;
-            req.Customer.Name = card.Name;
-            req.Customer.CountryId = card.CountryId;
-            req.Customer.CountryLevelId = card.CountryLevelId;
-            req.Customer.Address = card.Address;
-            req.Customer.Phone = card.Phone;
-            req.Customer.Email = card.Email;
-        }
-
-        req.Dte = FeFcv1Helper.Create(req, invoiceLines);
-
-        return req;
-    }
-
-    public static CreateDte01InvoiceRTCommand MockRtDte01()
-    {
-        var data = new
-        {
-            Branches = Db.Branches.ToList(),
-            Products = Db.Products.ToList(),
-            Persons = Db.Cards.Where(item => item.CardTypeId == (short)CardType.Person).ToList(),
-            WalkIns = Db.Cards.Where(item => item.CardTypeId == (short)CardType.WalkIn).ToList()
-        };
-
-        var lines = Random.Shared.Next(1, data.Products.Count);
-        var invoiceLines = new List<InvoiceLine>();
-
-        for (var i = 0; i < lines; i++)
-        {
-            var randomIndex = Random.Shared.Next(0, lines - 1);
-            var product = data.Products[randomIndex];
-            if (invoiceLines.Any(item => item.Code == product.Code))
-                continue;
-
-            invoiceLines.Add(new(product.Code, product.Description, (double)product.Price, 1));
-        }
-
-        var total = Math.Round(invoiceLines.Sum(item => item.Total), 2);
-
-        var req = new CreateDte01InvoiceRTCommand
-        {
-            PosId = (short)Random.Shared.Next(1, data.Branches.Count),
-            InvoiceTypeId = FeFcv1.TypeId,
-            InvoiceNumber = Convert.ToInt64($"{DateTime.Now:MMddhhmmss}"),
-            InvoiceDate = DateTime.Now,
-            InvoiceTotal = (decimal)total,
-            Lines = lines,
-        };
-
-        if (req.InvoiceTotal > CreateDte01InvoiceCommand.MaxAmountForAnonymousCustomers)
-        {
-            var persons = data.Persons.ToList();
-            var randomCard = persons[Random.Shared.Next(0, persons.Count - 1)];
-
-            req.Customer.Id = randomCard.Id;
-            req.Customer.DocumentTypeId = randomCard.DocumentTypeId;
-            req.Customer.DocumentNumber = randomCard.DocumentNumber;
-            req.Customer.WtId = randomCard.WtId;
-            req.Customer.Name = randomCard.Name;
-            req.Customer.CountryId = randomCard.CountryId;
-            req.Customer.CountryLevelId = randomCard.CountryLevelId;
-            req.Customer.Address = randomCard.Address;
-            req.Customer.Phone = randomCard.Phone;
-            req.Customer.Email = randomCard.Email;
-        }
-        else
-        {
-            var card = data.WalkIns.First();
-
-            req.Customer.Id = card.Id;
-            req.Customer.Name = card.Name;
-            req.Customer.CountryId = card.CountryId;
-            req.Customer.CountryLevelId = card.CountryLevelId;
-            req.Customer.Address = card.Address;
-            req.Customer.Phone = card.Phone;
-            req.Customer.Email = card.Email;
-        }
-
-        req.Dte = FeFcv1Helper.Create(req, invoiceLines);
-
-        return req;
-    }
-
-    public static CreateDte03InvoiceOWCommand MockOwDte03()
-    {
-        var data = new
-        {
-            Branches = Db.Branches.ToList(),
-            Products = Db.Products.ToList(),
-            Persons = Db.Cards.Where(item => item.CardTypeId == (short)CardType.Person).ToList()
-        };
-
-        var lines = Random.Shared.Next(1, data.Products.Count);
-        var invoiceLines = new List<InvoiceLine>();
-
-        for (var i = 0; i < lines; i++)
-        {
-            var randomIndex = Random.Shared.Next(0, lines - 1);
-            var product = data.Products[randomIndex];
-            if (invoiceLines.Any(item => item.Code == product.Code))
-                continue;
-
-            invoiceLines.Add(new(product.Code, product.Description, (double)product.Price, 1));
-        }
-
-        var total = Math.Round(invoiceLines.Sum(item => item.Total), 2);
-
-        var req = new CreateDte03InvoiceOWCommand
-        {
-            PosId = (short)Random.Shared.Next(1, data.Branches.Count),
-            InvoiceTypeId = FeCcfv3.TypeId,
-            InvoiceNumber = Convert.ToInt64($"{DateTime.Now:MMddhhmmss}"),
-            InvoiceDate = DateTime.Now,
-            InvoiceTotal = (decimal)total,
-            Lines = lines,
-        };
-
-        var persons = data.Persons.ToList();
-        var randomCard = persons[Random.Shared.Next(0, persons.Count - 1)];
-
-        req.Customer.Id = randomCard.Id;
-        req.Customer.DocumentTypeId = randomCard.DocumentTypeId;
-        req.Customer.DocumentNumber = randomCard.DocumentNumber;
-        req.Customer.WtId = randomCard.WtId;
-        req.Customer.Name = randomCard.Name;
-        req.Customer.CountryId = randomCard.CountryId;
-        req.Customer.CountryLevelId = randomCard.CountryLevelId;
-        req.Customer.Address = randomCard.Address;
-        req.Customer.Phone = randomCard.Phone;
-        req.Customer.Email = randomCard.Email;
-
-        req.Dte = FeCcfv3Helper.Create(req, invoiceLines);
-
-        return req;
-    }
-
-    public static CreateDte03InvoiceRTCommand MockRtDte03()
-    {
-        var data = new
-        {
-            Branches = Db.Branches.ToList(),
-            Products = Db.Products.ToList(),
-            Persons = Db.Cards.Where(item => item.CardTypeId == (short)CardType.Person).ToList()
-        };
-
-        var lines = Random.Shared.Next(1, data.Products.Count);
-        var invoiceLines = new List<InvoiceLine>();
-
-        for (var i = 0; i < lines; i++)
-        {
-            var randomIndex = Random.Shared.Next(0, lines - 1);
-            var product = data.Products[randomIndex];
-            if (invoiceLines.Any(item => item.Code == product.Code))
-                continue;
-
-            invoiceLines.Add(new(product.Code, product.Description, (double)product.Price, 1));
-        }
-
-        var total = Math.Round(invoiceLines.Sum(item => item.Total), 2);
-
-        var req = new CreateDte03InvoiceRTCommand
-        {
-            PosId = (short)Random.Shared.Next(1, data.Branches.Count),
-            InvoiceTypeId = FeCcfv3.TypeId,
-            InvoiceNumber = Convert.ToInt64($"{DateTime.Now:MMddhhmmss}"),
-            InvoiceDate = DateTime.Now,
-            InvoiceTotal = (decimal)total,
-            Lines = lines,
-        };
-
-        var persons = data.Persons.ToList();
-        var randomPerson = persons[Random.Shared.Next(0, persons.Count - 1)];
-
-        req.Customer.Id = randomPerson.Id;
-        req.Customer.DocumentTypeId = randomPerson.DocumentTypeId;
-        req.Customer.DocumentNumber = randomPerson.DocumentNumber;
-        req.Customer.WtId = randomPerson.WtId;
-        req.Customer.Name = randomPerson.Name;
-        req.Customer.CountryId = randomPerson.CountryId;
-        req.Customer.CountryLevelId = randomPerson.CountryLevelId;
-        req.Customer.Address = randomPerson.Address;
-        req.Customer.Phone = randomPerson.Phone;
-        req.Customer.Email = randomPerson.Email;
-
-        req.Dte = FeCcfv3Helper.Create(req, invoiceLines);
-
-        return req;
-    }
 }
